changed text-object, so they can be disabled...
git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@176 49b2a0ca-c541-4d49-a7ef-d785e4371512

+++ src/Geos/ripple.cpp
@@ -405,8 +405,8 @@ void ripple :: ripple_grab()
     
     if (index < RIPPLE_COUNT)
     {
-      cx[index] = 1.0*ctrX/xsize*GRID_SIZE_X;
-      cy[index] = 1.0*ctrY/ysize*GRID_SIZE_Y;
+      cx[index] = (int)(1.0*ctrX/xsize*GRID_SIZE_X);
+      cy[index] = (int)(1.0*ctrY/ysize*GRID_SIZE_Y);
       t[index] = 4*RIPPLE_STEP;
       max[index] = ripple_max_distance(cx[index], cy[index]);
